refactor: split giant components & add unoptimized comments
- Split gallery-main.tsx: 617→116 lines (hooks + GalleryGrid/GalleryLightbox)
- Split stats-dashboard-main.tsx: 617→200 lines (6 card components)
- Split moment-detail-main.tsx: 589→273 lines (hooks + CommentList/CommentInput)
- Split photo-viewer-main.tsx: 954→344 lines (4 hooks + 4 components)
- Split search.tsx: extracted search-main.tsx 163 lines

Item 5 completed:
- Add unoptimized comments to 8 files explaining why needed
- Update blog-optimization-fix-plan.md with completion status
