+++ src/components/admin/AdminImage.tsx
@@ -151,6 +151,7 @@ export function AdminImage({
           onError={handleError}
           onLoad={handleLoad}
           priority={priority}
+          // unoptimized: Skip Next.js optimization for external URLs not in remotePatterns
           unoptimized={!shouldOptimize}
           sizes={sizes}
         />
