+++ src/components/photo-viewer/components/index.ts
@@ -1,0 +1,4 @@
+export { PhotoViewerToolbar } from "./photo-viewer-toolbar";
+export { PhotoViewerNavigation } from "./photo-viewer-navigation";
+export { PhotoViewerImage } from "./photo-viewer-image";
+export { LoadingProgress } from "./loading-progress";
