+++ src/components/photo-viewer/types.ts
@@ -7,14 +7,16 @@ export type PhotoViewerProps = {
     prevPath?: string;
     nextPath?: string;
     locale?: "zh" | "en";
-    thumbnails?: {
-        id: string;
-        filePath: string;
-        microThumbPath?: string | null;
-        smallThumbPath?: string | null;
-        mediumPath?: string | null;
-    }[];
+    thumbnails?: ThumbnailItem[];
     currentId?: string;
+};
+
+export type ThumbnailItem = {
+    id: string;
+    filePath: string;
+    microThumbPath?: string | null;
+    smallThumbPath?: string | null;
+    mediumPath?: string | null;
 };
 
 export type OriginalLoadState = {
@@ -38,4 +40,84 @@ export type DragState = {
     originY: number;
 };
 
+export type Offset = { x: number; y: number };
+export type Size = { w: number; h: number };
+
+// ─────────────────────────────────────────────────────────────────────────────
+// Hook Return Types
+// ─────────────────────────────────────────────────────────────────────────────
+
+export type UsePhotoZoomReturn = {
+    scale: number;
+    setScale: React.Dispatch<React.SetStateAction<number>>;
+    offset: Offset;
+    setOffset: React.Dispatch<React.SetStateAction<Offset>>;
+    showZoomIndicator: boolean;
+    showHint: boolean;
+    setShowHint: (v: boolean) => void;
+    clampOffset: (nextOffset: Offset, s: number) => Offset;
+    containerSize: Size;
+    naturalSize: Size | null;
+    setNaturalSize: (size: Size | null) => void;
+    imgWrapRef: React.RefObject<HTMLDivElement | null>;
+};
+
+export type UsePhotoDragReturn = {
+    isDragging: boolean;
+    dragRef: React.MutableRefObject<DragState>;
+};
+
+export type UseImageLoadingReturn = {
+    displaySrc: string;
+    setDisplaySrc: (src: string) => void;
+    originalState: OriginalLoadState;
+    showProgress: boolean;
+};
+
+export type UseSlideAnimationReturn = {
+    slideContext: SlideContext | null;
+    startSlide: (direction: "left" | "right", from: { src: string; alt: string }) => void;
+    markPendingDirection: (direction: "prev" | "next") => void;
+    clearStoredDirection: () => void;
+    pendingDirectionRef: React.MutableRefObject<"prev" | "next" | null>;
+};
+
+// ─────────────────────────────────────────────────────────────────────────────
+// Component Props
+// ─────────────────────────────────────────────────────────────────────────────
+
+export type PhotoViewerToolbarProps = {
+    locale: "zh" | "en";
+    backButtonRef?: React.RefObject<HTMLAnchorElement | null>;
+};
+
+export type PhotoViewerNavigationProps = {
+    locale: "zh" | "en";
+    prevId: string | null;
+    nextId: string | null;
+    onPrevClick?: () => void;
+    onNextClick?: () => void;
+};
+
+export type PhotoViewerImageProps = {
+    locale: "zh" | "en";
+    displaySrc: string;
+    title?: string | null;
+    scale: number;
+    offset: Offset;
+    isDragging: boolean;
+    showZoomIndicator: boolean;
+    showHint: boolean;
+    slideContext: SlideContext | null;
+    imgWrapRef: React.RefObject<HTMLDivElement | null>;
+    onNaturalSizeChange: (size: Size) => void;
+};
+
+export type LoadingProgressProps = {
+    locale: "zh" | "en";
+    loadedBytes: number;
+    totalBytes: number | null;
+};
+
 export const SLIDE_STORAGE_KEY = "gallery-slide-direction";
+
