+++ src/components/search/components/index.ts
@@ -1,0 +1,5 @@
+export { SearchInput } from "./search-input";
+export { SearchHistoryList } from "./search-history";
+export { SearchResultsContainer } from "./search-results";
+export { MobileDrawer } from "./mobile-drawer";
+export { DesktopOverlay } from "./desktop-overlay";
