+++ src/components/search/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useSearch } from "./use-search";
+export { useSearchPosition } from "./use-search-position";
