+++ src/components/zhi/moment-detail/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useComments } from "./use-comments";
+export { useImageCarousel } from "./use-image-carousel";
